+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New Components
 - ScriptInput that uses a script to generate a list of Items
+- BulkFilter a new type of filter that does pre-processing with the list of Items for filters that need to handle all of the Items at once for efficiency purposes.
+- ScriptFilter that uses a script to validate items
 
 ### Bugs
 - Fix ScriptCommand to return when run_on_changes is set to true for batch runs and no changes are present.